# docs/openapi.yaml
openapi: 3.1.0
info:
  title: WikiMind Gateway
  description: Local LLM Knowledge OS — Personal API
  version: 0.1.0
paths:
...
  /api/wiki/contradictions:
    get:
      tags:
      - Wiki
      summary: List Contradictions
      description: List all contradictions for the user, optionally filtered by status.
      operationId: list_contradictions_api_wiki_contradictions_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ContradictionStatus'
          - type: 'null'
          title: Status
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContradictionResponse'
                title: Response List Contradictions Api Wiki Contradictions Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/contradictions/{contradiction_id}:
    get:
      tags:
      - Wiki
      summary: Get Contradiction
      description: Get a single contradiction with article details.
      operationId: get_contradiction_api_wiki_contradictions__contradiction_id__get
      parameters:
      - name: contradiction_id
        in: path
        required: true
        schema:
          type: string
          title: Contradiction Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContradictionResponse'
        '404':
          description: Contradiction not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Wiki
      summary: Resolve Persisted Contradiction
      description: Resolve or dismiss a contradiction.
      operationId: resolve_persisted_contradiction_api_wiki_contradictions__contradiction_id__patch
      parameters:
      - name: contradiction_id
        in: path
        required: true
        schema:
          type: string
          title: Contradiction Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveContradictionBody'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContradictionResponse'
        '404':
          description: Contradiction not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/articles:
    get:
      tags:
      - Wiki
      summary: List Articles
      description: List wiki articles with optional filtering and source provenance.
      operationId: list_articles_api_wiki_articles_get
      parameters:
      - name: concept
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Concept
      - name: confidence
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Confidence
      - name: page_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Page Type
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 50
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ArticleSummaryResponse'
                title: Response List Articles Api Wiki Articles Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/articles/stub:
    post:
      tags:
      - Wiki
      summary: Create Stub Article
      description: Create a stub wiki article — a placeholder page for a concept not
        yet compiled.
      operationId: create_stub_article_api_wiki_articles_stub_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStubRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateStubResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/wikilinks/resolve:
    get:
      tags:
      - Wiki
      summary: Resolve Wikilinks
      description: Search articles by partial title for wikilink autocomplete.
      operationId: resolve_wikilinks_api_wiki_wikilinks_resolve_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 1
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 50
          minimum: 1
          default: 10
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WikilinkMatch'
                title: Response Resolve Wikilinks Api Wiki Wikilinks Resolve Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/articles/random:
    get:
      tags:
      - Wiki
      summary: Get Random Article
      description: Return a random article belonging to the current user.
      operationId: get_random_article_api_wiki_articles_random_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleSummaryResponse'
        '404':
          description: No articles found
  /api/wiki/articles/{id_or_slug}:
    get:
      tags:
      - Wiki
      summary: Get Article
      description: Get full article by ID or slug, with content, backlinks, and source
        provenance.
      operationId: get_article_api_wiki_articles__id_or_slug__get
      parameters:
      - name: id_or_slug
        in: path
        required: true
        schema:
          type: string
          title: Id Or Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          description: Article not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - Wiki
      summary: Edit Article
      description: 'Manually edit an article''s content and/or title.


        Sets ``manually_edited=True`` so that recompilation respects user edits.

        Only the article owner can edit.'
      operationId: edit_article_api_wiki_articles__id_or_slug__patch
      parameters:
      - name: id_or_slug
        in: path
        required: true
        schema:
          type: string
          title: Id Or Slug
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArticleEditRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleResponse'
        '404':
          description: Article not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/graph:
    get:
      tags:
      - Wiki
      summary: Get Graph
      description: 'Full knowledge graph -- nodes and edges.


        Optional query parameters filter the returned edge set:


        * ``relation_type`` — keep only edges of one semantic relation type.

        * ``from_article`` — id or slug of the source article.

        * ``to_article`` — id or slug of the target article.


        Filters compose with AND. Filtering is pushed down into SQL.'
      operationId: get_graph_api_wiki_graph_get
      parameters:
      - name: relation_type
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/RelationType'
          - type: 'null'
          title: Relation Type
      - name: from_article
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: From Article
      - name: to_article
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: To Article
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/articles/{id_or_slug}/relationships:
    get:
      tags:
      - Wiki
      summary: Get Article Relationships
      description: Return typed relationships for an article, grouped by direction
        and type.
      operationId: get_article_relationships_api_wiki_articles__id_or_slug__relationships_get
      parameters:
      - name: id_or_slug
        in: path
        required: true
        schema:
          type: string
          title: Id Or Slug
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArticleRelationshipsResponse'
        '404':
          description: Article not found
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/search:
    get:
      tags:
      - Wiki
      summary: Search
      description: 'Full-text search across wiki articles with optional facet filters.


        Supports filtering by source_kind, page_type, concept, tag,

        date_range (7d, 30d, 365d), and staleness (low, medium, high).

        Sort by ``relevance`` (default) or ``recency``.'
      operationId: search_api_wiki_search_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: source_kind
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Source Kind
      - name: page_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Page Type
      - name: concept
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Concept
      - name: tag
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Tag
      - name: date_range
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Date Range
      - name: staleness
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Staleness
      - name: sort
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Sort
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/search/facets:
    get:
      tags:
      - Wiki
      summary: Search Facets
      description: 'Compute facet counts for the current search query.


        Returns counts grouped by: page_type, source_kind, concept,

        tag, date, and staleness.'
      operationId: search_facets_api_wiki_search_facets_get
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          title: Q
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FacetResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/wiki/concepts:
    get:
      tags:
      - Wiki
      summary: Get Concepts
      description: Concept taxonomy tree.
      operationId: get_concepts_api_wiki_concepts_get
      parameters:
      - name: include_empty
        in: query
        required: false
        schema:
          type: boolean
          default: true
          title: Include Empty
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    ContradictionResponse:
      properties:
        id:
          type: string
          title: Id
        claim_a:
          type: string
          title: Claim A
        claim_b:
          type: string
          title: Claim B
        article_a_id:
          type: string
          title: Article A Id
        article_b_id:
          type: string
          title: Article B Id
        article_a_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Article A Title
        article_b_title:
          anyOf:
          - type: string
          - type: 'null'
          title: Article B Title
        detected_at:
          type: string
          format: date-time
          title: Detected At
        status:
          type: string
          title: Status
        resolution:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolution
        resolved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Resolved At
        resolved_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Resolved By
      type: object
      required:
      - id
      - claim_a
      - claim_b
      - article_a_id
      - article_b_id
      - detected_at
      - status
      title: ContradictionResponse
      description: API response for a single persisted contradiction.
    ContradictionStatus:
      type: string
      enum:
      - active
      - resolved
      - dismissed
      title: ContradictionStatus
      description: Lifecycle status of a persisted contradiction.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...