# src/wikimind/api/routes/wiki.py
"""Endpoints for browsing wiki articles, knowledge graph, and search."""

import structlog
from fastapi import APIRouter, Depends, Query
from sqlmodel.ext.asyncio.session import AsyncSession

from wikimind.api.deps import get_current_user_id, require_plan
from wikimind.database import get_session
from wikimind.models import (
    ArticleEditRequest,
    ArticleRelationshipsResponse,
    ArticleResponse,
    ArticleSummaryResponse,
    ArticleTagResponse,
    ContradictionResolution,
    ContradictionResolutionOption,
    ContradictionResponse,
    ContradictionStatus,
    CreateStubRequest,
    CreateStubResponse,
    FacetResponse,
    GraphResponse,
    HealthSummaryResponse,
    Plan,
    RebuildConceptsResponse,
    RecompileResponse,
    RefreshArticleResponse,
    RelationType,
    ResolveContradictionBody,
    ResolveContradictionRequest,
    ResolveContradictionResponse,
    SearchResponse,
    SearchResult,
    TagArticleRequest,
    TagResponse,
    WikilinkMatch,
)
from wikimind.services.contradiction import ContradictionService
from wikimind.services.factories import (
    get_contradiction_service,
    get_linter_service,
    get_search_service,
    get_tag_service,
    get_wiki_service,
)
from wikimind.services.linter import LinterService
from wikimind.services.quota import check_article_quota
from wikimind.services.search import SearchService
from wikimind.services.tags import TagService
from wikimind.services.taxonomy import rebuild_taxonomy
from wikimind.services.wiki import WikiService, _staleness_score

log = structlog.get_logger()

router = APIRouter()


@router.get(
    "/contradiction-resolutions",
    response_model=list[ContradictionResolutionOption],
)
async def list_contradiction_resolutions():
    """Return the valid contradiction resolution options."""
    return [
        ContradictionResolutionOption(value=r.value, label=r.value.replace("_", " ").title())
        for r in ContradictionResolution
    ]


@router.get("/contradictions", response_model=list[ContradictionResponse])
async def list_contradictions(
    status: ContradictionStatus | None = None,
    limit: int = Query(default=50, ge=1, le=200),
    offset: int = Query(default=0, ge=0),
    session: AsyncSession = Depends(get_session),
    service: ContradictionService = Depends(get_contradiction_service),
    user_id: str = Depends(get_current_user_id),
):
    """List all contradictions for the user, optionally filtered by status."""
    return await service.list_contradictions(session, user_id=user_id, status=status, limit=limit, offset=offset)


@router.get(
    "/contradictions/{contradiction_id}",
    response_model=ContradictionResponse,
    responses={404: {"description": "Contradiction not found"}},
)
async def get_contradiction(
    contradiction_id: str,
    session: AsyncSession = Depends(get_session),
    service: ContradictionService = Depends(get_contradiction_service),
    user_id: str = Depends(get_current_user_id),
):
    """Get a single contradiction with article details."""
    return await service.get_contradiction(session, contradiction_id, user_id=user_id)


@router.patch(
    "/contradictions/{contradiction_id}",
    response_model=ContradictionResponse,
    responses={404: {"description": "Contradiction not found"}},
)
async def resolve_persisted_contradiction(
    contradiction_id: str,
    body: ResolveContradictionBody,
    session: AsyncSession = Depends(get_session),
    service: ContradictionService = Depends(get_contradiction_service),
    user_id: str = Depends(get_current_user_id),
):
    """Resolve or dismiss a contradiction."""
    return await service.resolve_contradiction(
        session,
        contradiction_id,
        user_id=user_id,
        new_status=body.status,
        resolution=body.resolution,
    )


@router.get("/articles", response_model=list[ArticleSummaryResponse])
async def list_articles(
    concept: str | None = None,
    confidence: str | None = None,
    page_type: str | None = None,
    limit: int = Query(default=50, ge=1, le=200),
    offset: int = Query(default=0, ge=0),
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """List wiki articles with optional filtering and source provenance."""
    return await service.list_articles(
        session,
        concept=concept,
        confidence=confidence,
        page_type=page_type,
        limit=limit,
        offset=offset,
        user_id=user_id,
    )


@router.post(
    "/articles/stub",
    response_model=CreateStubResponse,
    status_code=201,
)
async def create_stub_article(
    body: CreateStubRequest,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
    plan: Plan | None = Depends(require_plan),
):
    """Create a stub wiki article — a placeholder page for a concept not yet compiled."""
    if plan:
        await check_article_quota(session, user_id, plan)
    return await service.create_stub_article(
        title=body.title,
        body_markdown=body.body_markdown,
        session=session,
        user_id=user_id,
    )


@router.get(
    "/wikilinks/resolve",
    response_model=list[WikilinkMatch],
)
async def resolve_wikilinks(
    q: str = Query(..., min_length=1),
    limit: int = Query(10, ge=1, le=50),
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Search articles by partial title for wikilink autocomplete."""
    return await service.resolve_wikilinks(q, session, user_id=user_id, limit=limit)


@router.get(
    "/articles/random",
    response_model=ArticleSummaryResponse,
    responses={404: {"description": "No articles found"}},
)
async def get_random_article(
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Return a random article belonging to the current user."""
    return await service.get_random_article(session, user_id=user_id)


@router.get(
    "/articles/{id_or_slug}",
    response_model=ArticleResponse,
    responses={404: {"description": "Article not found"}},
)
async def get_article(
    id_or_slug: str,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Get full article by ID or slug, with content, backlinks, and source provenance."""
    return await service.get_article(id_or_slug, session, user_id=user_id)


@router.patch(
    "/articles/{id_or_slug}",
    response_model=ArticleResponse,
    responses={404: {"description": "Article not found"}},
)
async def edit_article(
    id_or_slug: str,
    body: ArticleEditRequest,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Manually edit an article's content and/or title.

    Sets ``manually_edited=True`` so that recompilation respects user edits.
    Only the article owner can edit.
    """
    return await service.edit_article(
        id_or_slug,
        session,
        user_id=user_id,
        content=body.content,
        title=body.title,
    )


@router.get("/graph", response_model=GraphResponse)
async def get_graph(
    relation_type: RelationType | None = Query(default=None),
    from_article: str | None = Query(default=None),
    to_article: str | None = Query(default=None),
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Full knowledge graph -- nodes and edges.

    Optional query parameters filter the returned edge set:

    * ``relation_type`` — keep only edges of one semantic relation type.
    * ``from_article`` — id or slug of the source article.
    * ``to_article`` — id or slug of the target article.

    Filters compose with AND. Filtering is pushed down into SQL.
    """
    return await service.get_graph(
        session,
        user_id=user_id,
        relation_type=relation_type,
        from_article=from_article,
        to_article=to_article,
    )


@router.get(
    "/articles/{id_or_slug}/relationships",
    response_model=ArticleRelationshipsResponse,
    responses={404: {"description": "Article not found"}},
)
async def get_article_relationships(
    id_or_slug: str,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Return typed relationships for an article, grouped by direction and type."""
    return await service.get_relationships(id_or_slug, session, user_id=user_id)


@router.get("/search", response_model=SearchResponse)
async def search(
    q: str = Query(..., min_length=2),
    limit: int = Query(20, ge=1, le=100),
    offset: int = Query(0, ge=0),
    source_kind: str | None = Query(None),
    page_type: str | None = Query(None),
    concept: str | None = Query(None),
    tag: str | None = Query(None),
    date_range: str | None = Query(None),
    staleness: str | None = Query(None),
    sort: str | None = Query(None),
    session: AsyncSession = Depends(get_session),
    search_service: SearchService = Depends(get_search_service),
    user_id: str = Depends(get_current_user_id),
):
    """Full-text search across wiki articles with optional facet filters.

    Supports filtering by source_kind, page_type, concept, tag,
    date_range (7d, 30d, 365d), and staleness (low, medium, high).
    Sort by ``relevance`` (default) or ``recency``.
    """
    fts_response = await search_service.search(
        q,
        session,
        user_id=user_id,
        limit=limit,
        offset=offset,
        source_kind=source_kind,
        page_type=page_type,
        concept=concept,
        tag=tag,
        date_range=date_range,
        staleness=staleness,
        sort=sort,
    )
    results = [
        SearchResult(
            article_id=r.article_id,
            slug=r.slug,
            title=r.title,
            snippet=r.snippet,
            rank=r.rank,
        )
        for r in fts_response.results
    ]
    return SearchResponse(results=results, total=fts_response.total, query=q)


@router.get("/search/facets", response_model=FacetResponse)
async def search_facets(
    q: str = Query(..., min_length=2),
    session: AsyncSession = Depends(get_session),
    search_service: SearchService = Depends(get_search_service),
    user_id: str = Depends(get_current_user_id),
):
    """Compute facet counts for the current search query.

    Returns counts grouped by: page_type, source_kind, concept,
    tag, date, and staleness.
    """
    return await search_service.get_facets(q, session, user_id=user_id)


@router.get("/concepts")
async def get_concepts(
    include_empty: bool = True,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Concept taxonomy tree."""
    return await service.get_concepts(session, include_empty=include_empty, user_id=user_id)


@router.post("/concepts/rebuild", response_model=RebuildConceptsResponse)
async def rebuild_concepts(
    session: AsyncSession = Depends(get_session),
    user_id: str = Depends(get_current_user_id),
):
    """Trigger LLM-powered taxonomy hierarchy rebuild."""
    await rebuild_taxonomy(session, user_id=user_id)
    return RebuildConceptsResponse(status="ok")


@router.get(
    "/concepts/{name}",
    responses={404: {"description": "Concept not found"}},
)
async def get_concept(
    name: str,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Concept detail with linked articles."""
    return await service.get_concept(name, session, user_id=user_id)


@router.get("/concepts/{name}/articles", response_model=list[ArticleSummaryResponse])
async def get_concept_articles(
    name: str,
    limit: int = Query(default=50, ge=1, le=200),
    offset: int = Query(default=0, ge=0),
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Articles tagged with this concept."""
    return await service.get_concept_articles(name, session, limit=limit, offset=offset, user_id=user_id)


@router.get("/health", response_model=HealthSummaryResponse)
async def get_health(
    session: AsyncSession = Depends(get_session),
    linter_service: LinterService = Depends(get_linter_service),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Latest wiki health report from linter.

    DEPRECATED: Use GET /lint/reports/latest instead. This endpoint
    delegates to the new LinterService for backward compatibility.
    """
    return await service.get_health_summary(session, user_id=user_id, linter_service=linter_service)


@router.post(
    "/backlinks/{source_id}/{target_id}/resolve",
    response_model=ResolveContradictionResponse,
    deprecated=True,
)
async def resolve_contradiction(
    source_id: str,
    target_id: str,
    body: ResolveContradictionRequest,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Resolve a contradiction between two articles.

    DEPRECATED: Use ``PATCH /wiki/contradictions/{id}`` instead. This endpoint
    updates the Backlink for backward compatibility AND forwards the resolution
    to the Contradiction table (single source of truth).
    """
    return await service.resolve_legacy_contradiction(
        source_id=source_id,
        target_id=target_id,
        resolution=body.resolution,
        resolution_note=body.resolution_note,
        session=session,
        user_id=user_id,
    )


@router.post(
    "/articles/{id_or_slug}/refresh",
    response_model=RefreshArticleResponse,
    responses={404: {"description": "Article not found"}},
)
async def refresh_article(
    id_or_slug: str,
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Mark an article as 'still current' without recompiling.

    Creates a manual_refresh reinforcement event and resets the article's
    staleness score. Use this when you have reviewed an article and
    confirmed its content is still accurate.
    """
    article = await service.refresh_article(id_or_slug, session, user_id=user_id)
    return RefreshArticleResponse(
        status="refreshed",
        staleness_score=_staleness_score(article),
    )


@router.post(
    "/articles/{article_id}/recompile",
    response_model=RecompileResponse,
    responses={409: {"description": "Article has manual edits"}},
)
async def recompile_article(
    article_id: str,
    mode: str | None = Query(default=None),
    force: bool = Query(default=False),
    session: AsyncSession = Depends(get_session),
    service: WikiService = Depends(get_wiki_service),
    user_id: str = Depends(get_current_user_id),
):
    """Schedule an async recompilation job for an article.

    If the article has been manually edited (``manually_edited=True``),
    returns 409 Conflict unless ``force=true`` is passed. When forced,
    the manual edits flag is cleared before recompilation.
    """
    return await service.recompile_article(
        article_id=article_id,
        session=session,
        user_id=user_id,
        mode=mode,
        force=force,
    )


# ---------------------------------------------------------------------------
# Article tagging endpoints
# ---------------------------------------------------------------------------


@router.post(
    "/articles/{article_id}/tags",
    response_model=ArticleTagResponse,
    status_code=201,
    responses={404: {"description": "Article or tag not found"}},
)
async def tag_article(
    article_id: str,
    body: TagArticleRequest,
    session: AsyncSession = Depends(get_session),
    tag_service: TagService = Depends(get_tag_service),
    user_id: str = Depends(get_current_user_id),
):
    """Apply a tag to an article."""
    await tag_service.tag_article(
        session,
        article_id=article_id,
        tag_id=body.tag_id,
        user_id=user_id,
    )
    return ArticleTagResponse(article_id=article_id, tag_id=body.tag_id)


@router.delete(
    "/articles/{article_id}/tags/{tag_id}",
    status_code=204,
    responses={404: {"description": "Tag association not found"}},
)
async def untag_article(
    article_id: str,
    tag_id: str,
    session: AsyncSession = Depends(get_session),
    tag_service: TagService = Depends(get_tag_service),
    user_id: str = Depends(get_current_user_id),
):
    """Remove a tag from an article."""
    await tag_service.untag_article(
        session,
        article_id=article_id,
        tag_id=tag_id,
        user_id=user_id,
    )


@router.get(
    "/articles/{article_id}/tags",
    response_model=list[TagResponse],
    responses={404: {"description": "Article not found"}},
)
async def get_article_tags(
    article_id: str,
    session: AsyncSession = Depends(get_session),
    tag_service: TagService = Depends(get_tag_service),
    user_id: str = Depends(get_current_user_id),
):
    """Get all tags applied to an article."""
    return await tag_service.get_tags_for_article(session, article_id=article_id, user_id=user_id)